Fix SystemTap dtrace warning for smgr probe argument types
Commits ca326e903d and 1f8c504e308 widened the byte-count arguments of
smgr__md__read__done and smgr__md__write__done to "long long int".
SystemTap's dtrace(1) fails on that specific spelling and falls back
with a warning (misreported near the previous probe).  Use ssize_t and
size_t instead, matching the md.c call sites.  Revise the probes.d
note: prefer system-supplied type names (macOS dtrace rejects names
like PostgreSQL's uint64), and call out "long long int" as a known
SystemTap failure rather than a blanket multi-word ban.

Author: Andrey Rachitskiy <pl0h0yp1@gmail.com>
Reported-by: Laurenz Albe <laurenz.albe@cybertec.at>
Discussion: https://postgr.es/m/697d3c88568442cd637d8453d129f1bb14bbd2a8.camel@cybertec.at

+++ src/backend/utils/probes.d
@@ -11,8 +11,11 @@
 /*
  * Typedefs used in PostgreSQL probes.
  *
- * NOTE: Do not use system-provided typedefs (e.g. uintptr_t, uint32_t, etc)
- * in probe definitions, as they cause compilation errors on macOS.
+ * NOTE: Prefer system-supplied type names in probe definitions (e.g.
+ * uint64_t, not PostgreSQL's uint64); macOS' dtrace rejects the latter.
+ * Also avoid "long long int", as SystemTap's dtrace(1) fails on that
+ * specific spelling.  PG type aliases below must be #define'd so that
+ * cpp expands them before dtrace sees the file.
  */
 #define LocalTransactionId unsigned int
 #define LWLockMode int
@@ -83,9 +86,9 @@ provider postgresql {
 	probe twophase__checkpoint__done();
 
 	probe smgr__md__read__start(ForkNumber, BlockNumber, Oid, Oid, Oid, int);
-	probe smgr__md__read__done(ForkNumber, BlockNumber, Oid, Oid, Oid, int, long long int, long long int);
+	probe smgr__md__read__done(ForkNumber, BlockNumber, Oid, Oid, Oid, int, ssize_t, size_t);
 	probe smgr__md__write__start(ForkNumber, BlockNumber, Oid, Oid, Oid, int);
-	probe smgr__md__write__done(ForkNumber, BlockNumber, Oid, Oid, Oid, int, long long int, long long int);
+	probe smgr__md__write__done(ForkNumber, BlockNumber, Oid, Oid, Oid, int, ssize_t, size_t);
 
 	probe wal__insert(unsigned char, unsigned char);
 	probe wal__switch();
